fix(arborist): honor allowRemote exemption for registry bundle tarballs under --dry-run/--package-lock-only (#9800)
